+++ libs/artifact/src/vivarium/artifact/__init__.py
@@ -1,4 +1,7 @@
-"""Vivarium data artifacts.
+"""
+=================
+Vivarium Artifact
+=================
 
 A data artifact is an archive on disk that packages all data relevant to a
 particular simulation. This package provides the artifact model and the HDF5